+++ src/SDK/Client/Client.ts
@@ -1,15 +1,20 @@
+import { EventEmitter } from 'events';
 import { Account, Wallet } from "@dashevo/wallet-lib";
 import DAPIClientTransport from "@dashevo/wallet-lib/src/transport/DAPIClientTransport/DAPIClientTransport"
 import { Platform } from './Platform';
 import { Network } from "@dashevo/dashcore-lib";
 import DAPIClient from "@dashevo/dapi-client";
+import { ClientApps, ClientAppsOptions } from "./ClientApps";
+
+export interface WalletOptions extends Wallet.IWalletOptions {
+    defaultAccountIndex?: number;
+}
 
 /**
  * Interface Client Options
  *
  * @param {ClientApps?} [apps] - applications
- * @param {Wallet.IWalletOptions} [wallet] - Wallet options
- * @param {number} [walletAccountIndex=0] - Wallet account index number
+ * @param {WalletOptions} [wallet] - Wallet options
  * @param {DAPIAddressProvider} [dapiAddressProvider] - DAPI Address Provider instance
  * @param {Array<RawDAPIAddress|DAPIAddress|string>} [dapiAddresses] - DAPI addresses
  * @param {string[]|RawDAPIAddress[]} [seeds] - DAPI seeds
@@ -19,9 +24,8 @@ import DAPIClient from "@dashevo/dapi-client";
  * @param {number} [baseBanTime=60000]
  */
 export interface ClientOpts {
-    apps?: ClientApps,
-    wallet?: Wallet.IWalletOptions,
-    walletAccountIndex?: number,
+    apps?: ClientAppsOptions,
+    wallet?: WalletOptions,
     dapiAddressProvider?: any,
     dapiAddresses?: any[],
     seeds?: any[],
@@ -29,28 +33,19 @@ export interface ClientOpts {
     timeout?: number,
     retries?: number,
     baseBanTime?: number,
-}
-
-/**
- * Interface for ClientApps
- */
-export interface ClientApps {
-    [name: string]: {
-        contractId: string,
-        contract?: any
-    }
+    driveProtocolVersion?: number,
 }
 
 /**
  * Client class that wraps all components together to allow integrated payments on both the Dash Network (layer 1)
  * and the Dash Platform (layer 2).
  */
-export class Client {
-    public network: string = 'evonet';
+export class Client extends EventEmitter {
+    public network: string = 'testnet';
     public wallet: Wallet | undefined;
     public account: Account | undefined;
-    public platform: Platform | undefined;
-    public walletAccountIndex: number | undefined = 0;
+    public platform: Platform;
+    public defaultAccountIndex: number | undefined = 0;
     private readonly dapiClient: DAPIClient;
     private readonly apps: ClientApps;
     private options: ClientOpts;
@@ -61,12 +56,11 @@ export class Client {
      * @param {ClientOpts} [options] - options for SDK Client
      */
     constructor(options: ClientOpts = {}) {
-        this.options = {
-            walletAccountIndex: 0,
-            ...options
-        }
+        super();
 
-        this.network = this.options.network ? this.options.network.toString() : 'evonet';
+        this.options = options;
+
+        this.network = this.options.network ? this.options.network.toString() : 'testnet';
 
         // Initialize DAPI Client
         const dapiClientOptions = {
@@ -103,18 +97,27 @@ export class Client {
             });
 
             // @ts-ignore
-            this.walletAccountIndex = this.options.walletAccountIndex;
+            this.wallet.on('error', (error, context) => (
+                this.emit('error', error, { wallet: context })
+            ));
         }
 
-        this.apps = Object.assign({
+        // @ts-ignore
+        this.defaultAccountIndex = this.options.wallet?.defaultAccountIndex || 0;
+
+        this.apps = new ClientApps(Object.assign({
             dpns: {
-                contractId: '7DVe2cDyZMf8sDjQ46XqDzbeGKncrmkD6L96QohLmLbg'
-            }
-        }, this.options.apps);
+                contractId: 'Bw9PUC3aSEGQ4j5qrvpNLrRNFPVMiUHZLr1atgfYJcmf'
+            },
+            dashpay: {
+                contractId: '2Vuou3EfbrtunwCZvQp1XS5PXZ5CgC1pGBz4VPT4ojmy',
+            },
+        }, this.options.apps));
 
         this.platform = new Platform({
             client: this,
-            apps: this.getApps(),
+            network: this.network,
+            driveProtocolVersion: this.options.driveProtocolVersion,
         });
     }
 
@@ -130,7 +133,7 @@ export class Client {
         }
 
         options = {
-            index: this.walletAccountIndex,
+            index: this.defaultAccountIndex,
             ...options,
         }