+++ src/SDK/Client/Platform/methods/contracts/broadcast.ts
@@ -1,3 +1,7 @@
+import { Platform } from "../../Platform";
+import broadcastStateTransition from "../../broadcastStateTransition";
+import { signStateTransition } from "../../signStateTransition";
+
 /**
  * Broadcast contract onto the platform
  *
@@ -6,15 +10,15 @@
  * @param identity - identity
  * @return dataContract
  */
-import { Platform } from "../../Platform";
-import broadcastStateTransition from "../../broadcastStateTransition";
-
 export default async function broadcast(this: Platform, dataContract: any, identity: any): Promise<any> {
+    await this.initialize();
+
     const { dpp } = this;
 
     const dataContractCreateTransition = dpp.dataContract.createStateTransition(dataContract);
 
-    await broadcastStateTransition(this, dataContractCreateTransition, identity);
+    await signStateTransition(this, dataContractCreateTransition, identity);
+    await broadcastStateTransition(this, dataContractCreateTransition);
 
-    return dataContract;
+    return dataContractCreateTransition;
 }