+++ src/SDK/Client/Platform/methods/documents/create.ts
@@ -13,18 +13,22 @@ declare interface createOpts {
  * @param {Object} [data] - options
  */
 export async function create(this: Platform, typeLocator: string, identity: any, data: createOpts = {}): Promise<any> {
+    await this.initialize();
+
     const { dpp } = this;
 
-    const appNames = Object.keys(this.apps);
+    const appNames = this.client.getApps().getNames();
+
     //We can either provide of type `dashpay.profile` or if only one schema provided, of type `profile`.
     const [appName, fieldType] = (typeLocator.includes('.')) ? typeLocator.split('.') : [appNames[0], typeLocator];
 
+    const { contractId } = this.client.getApps().get(appName);
 
-    if(!this.apps[appName]){
-        throw new Error(`Cannot find contractId for ${appName}`);
+    const dataContract = await this.contracts.get(contractId);
+
+    if (dataContract === null) {
+        throw new Error(`Contract ${appName} not found. Ensure contractId ${contractId} is correct.`)
     }
-
-    const dataContract = await this.contracts.get(this.apps[appName].contractId);
 
     return dpp.document.create(
         dataContract,