+++ src/SDK/Client/Platform/methods/identities/get.ts
@@ -1,21 +1,46 @@
 import {Platform} from "../../Platform";
+// @ts-ignore
+import Identifier from "@dashevo/dpp/lib/Identifier";
+import Metadata from "@dashevo/dpp/lib/Metadata";
+const NotFoundError = require('@dashevo/dapi-client/lib/transport/GrpcTransport/errors/NotFoundError');
 
 /**
  * Get an identity from the platform
  *
  * @param {Platform} this - bound instance class
- * @param {string} id - id
+ * @param {string|Identifier} id - id
  * @returns Identity
  */
-export async function get(this: Platform, id: string): Promise<any> {
-    // @ts-ignore
-    const identityBuffer = await this.client.getDAPIClient().platform.getIdentity(id);
+export async function get(this: Platform, id: Identifier|string): Promise<any> {
+    await this.initialize();
 
-    if (identityBuffer === null) {
-        return null;
+    const identifier = Identifier.from(id);
+
+    let identityResponse;
+    try {
+        identityResponse = await this.client.getDAPIClient().platform.getIdentity(identifier);
+    } catch (e) {
+        if (e instanceof NotFoundError) {
+            return null;
+        }
+
+        throw e;
     }
 
-    return this.dpp.identity.createFromSerialized(identityBuffer);
+    const identity = this.dpp.identity.createFromBuffer(identityResponse.getIdentity());
+
+    let metadata = null;
+    const responseMetadata = identityResponse.getMetadata();
+    if (responseMetadata) {
+        metadata = new Metadata({
+            blockHeight: responseMetadata.getHeight(),
+            coreChainLockedHeight: responseMetadata.getCoreChainLockedHeight(),
+        });
+    }
+
+    identity.setMetadata(metadata);
+
+    return identity;
 }
 
 export default get;