+++ src/SDK/Client/Platform/methods/names/search.ts
@@ -7,12 +7,15 @@ import {Platform} from "../../Platform";
  * @returns Documents[] - The array of documents that match the search parameters.
  */
 export async function search(this: Platform, labelPrefix: string, parentDomainName: string = '') {
+    await this.initialize();
+
     const normalizedParentDomainName = parentDomainName.toLowerCase();
+    const normalizedLabelPrefix = labelPrefix.toLowerCase();
 
     const documents = await this.documents.get('dpns.domain', {
         where: [
             ['normalizedParentDomainName', '==', normalizedParentDomainName],
-            ['normalizedLabel', 'startsWith', labelPrefix],
+            ['normalizedLabel', 'startsWith', normalizedLabelPrefix],
         ],
     });